lesson_flask_docker/pyproject.toml
[tool.poetry]
name = "Homework Flask&Docker"
version = "0.1.0"
description = ""
authors = ["Yuriy Peskov <yuriy.peskov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.7.0"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
